Removed class GOMidiReceiverBase
