+++ src/grandorgue/CMakeLists.txt
@@ -145,7 +145,6 @@ loader/cache/GOCacheCleaner.cpp
 loader/cache/GOCacheWriter.cpp
 midi/dialog-creator/GOMidiDialogCreatorProxy.cpp
 midi/elements/GOMidiReceiver.cpp
-midi/elements/GOMidiReceiverBase.cpp
 midi/elements/GOMidiShortcutReceiver.cpp
 midi/elements/GOMidiSendProxy.cpp
 midi/elements/GOMidiSender.cpp
