+++ src/grandorgue/GODocument.cpp
@@ -209,7 +209,7 @@ void GODocument::ShowMIDIEventDialog(
   void *element,
   const wxString &title,
   const wxString &dialogSelector,
-  GOMidiReceiverBase *event,
+  GOMidiReceiver *event,
   GOMidiSender *sender,
   GOMidiShortcutReceiver *key,
   GOMidiSender *division,
