+++ src/grandorgue/config/GORegisteredOrgan.cpp
@@ -10,8 +10,10 @@
 #include "config/GOConfigReader.h"
 #include "config/GOConfigWriter.h"
 
+#include "GOConfig.h"
+
 GORegisteredOrgan::GORegisteredOrgan(
-  GOConfigReader &cfg, const wxString &group, GOMidiMap &map)
+  GOConfig &config, GOConfigReader &cfg, const wxString &group)
   : GOOrgan(
     cfg.ReadString(CMBSetting, group, wxT("ODFPath")),
     cfg.ReadString(CMBSetting, group, wxT("Archiv"), false),
@@ -19,10 +21,10 @@ GORegisteredOrgan::GORegisteredOrgan(
     cfg.ReadString(CMBSetting, group, wxT("ChurchName")),
     cfg.ReadString(CMBSetting, group, wxT("OrganBuilder")),
     cfg.ReadString(CMBSetting, group, wxT("RecordingDetail"))),
-    m_midi(MIDI_RECV_ORGAN) {
+    m_midi(config, MIDI_RECV_ORGAN) {
   m_LastUse = cfg.ReadInteger(
     CMBSetting, group, wxT("LastUse"), 0, INT_MAX, false, m_LastUse);
-  m_midi.Load(cfg, group, map);
+  m_midi.Load(cfg, group, config.GetMidiMap());
 }
 
 void GORegisteredOrgan::Save(
