+++ src/grandorgue/gui/dialogs/settings/GOSettingsMidiMessage.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -52,7 +52,7 @@ GOSettingsMidiMessage::GOSettingsMidiMessage(
   topSizer->Add(m_Properties, 0, wxALIGN_RIGHT | wxALL, 5);
 
   for (unsigned i = 0; i < m_config.GetEventCount(); i++) {
-    const GOMidiReceiverBase *recv = m_config.GetMidiEvent(i);
+    const GOMidiReceiver *recv = m_config.GetMidiEvent(i);
 
     m_Events->InsertItem(i, m_config.GetEventGroup(i));
     m_Events->SetItemPtrData(i, (wxUIntPtr)recv);
@@ -77,8 +77,8 @@ void GOSettingsMidiMessage::OnEventsDoubleClick(wxListEvent &event) {
   m_Properties->Enable();
   int index = m_Events->GetFirstSelected();
 
-  GOMidiReceiverBase *recv
-    = (GOMidiReceiverBase *)m_Events->GetItemData(m_Events->GetFirstSelected());
+  GOMidiReceiver *recv
+    = (GOMidiReceiver *)m_Events->GetItemData(m_Events->GetFirstSelected());
   GOMidiEventDialog dlg(
     NULL,
     this,
