+++ src/grandorgue/midi/dialog-creator/GOMidiDialogCreatorProxy.h
@@ -23,7 +23,7 @@ public:
     void *element,
     const wxString &title,
     const wxString &dialogSelector,
-    GOMidiReceiverBase *event,
+    GOMidiReceiver *event,
     GOMidiSender *sender,
     GOMidiShortcutReceiver *key,
     GOMidiSender *division = nullptr,
